refactor(signals): move evaluateEventContextSignal to its own file
classifier_signal_context.go exceeded the 800-line hard limit enforced by the
structure checker. Moving evaluateEventContextSignal to a dedicated file
(classifier_signal_event_context.go) follows the same pattern used for
classifier_signal_structure.go, classifier_signal_conversation.go, etc.
